+++ src/Loader/Modes/Render.js
@@ -43,11 +43,12 @@ class Render extends PreprocessorMode {
    * Export template code with rendered HTML.
    *
    * @param {string} content The template content.
-   * @param {{}} data The object with variables passed in template.
-   * @param {string} issuer The issuer of the template file.
+   * @param {BundlerPluginLoaderContext} loaderContext
    * @return {string}
    */
-  export(content, { data, resource: issuer }) {
+  export(content, loaderContext) {
+    const { resource: issuer } = loaderContext;
+
     /* istanbul ignore next: Webpack API no provide `loaderContext.hot` for testing */
     if (this.hot && PluginService.useHotUpdate()) {
       content = this.injectHotScript(content);