+++ test/cases/js-tmpl-ejs-compile/src/app.js
@@ -1,7 +1,11 @@
 import tmpl from './partials/content.html?lang=en'; // test: pass variable into partials via query
 
 // render template function with variables
-const html = tmpl({ name: 'World', people: ['Alexa <Amazon>', 'Cortana <MS>', 'Siri <Apple>'] });
+const html = tmpl({
+  name: 'World',
+  people: ['Alexa <Amazon>', 'Cortana <MS>', 'Siri <Apple>'],
+  nested: { name: 'EJS' },
+});
 
 document.getElementById('main').innerHTML = html;