+++ test/cases/js-tmpl-eta-render-many-pages/src/app.js
@@ -1,0 +1,3 @@
+import './nested';
+
+console.log('>> app');