+++ test/cases/js-tmpl-eta-render-many-pages/src/content.js
@@ -1,0 +1,4 @@
+// pass variable name into template via query
+import html from './partials/content.html?render&name=World';
+
+export { html };