+++ test/cases/js-tmpl-eta-render-many-pages/src/nested.js
@@ -1,0 +1,3 @@
+import { html } from './content';
+
+document.getElementById('content').innerHTML = html;